+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## [Unreleased]
 
-- Add `buf beta registry repository update` command. So far only with support
-  to update visibility.
+- Add `buf beta registry repository update` command which supports
+  visibility for now.
 
 ## [v1.3.1] - 2022-03-30